kernel: move modules code to modules.h
