+++ src/gap.h
@@ -16,6 +16,7 @@
 #include <src/system.h>
 
 #include <src/error.h>
+#include <src/modules.h>
 
 /****************************************************************************
 **
@@ -84,175 +85,6 @@ enum {
     STATUS_ERROR       = 32,    // error
     STATUS_QQUIT       = 64,    // QUIT command
 };
-
-
-/****************************************************************************
-**
-*F * * * * * * * * * * * * * initialize package * * * * * * * * * * * * * * *
-*/
-
-
-/****************************************************************************
-**
-*F  InitBagNamesFromTable( <table> )  . . . . . . . . .  initialise bag names
-*/
-extern void InitBagNamesFromTable(const StructBagNames * tab);
-
-
-/****************************************************************************
-**
-*F  InitClearFiltsTNumsFromTable( <tab> ) . . .  initialise clear filts tnums
-*/
-extern void InitClearFiltsTNumsFromTable(const Int * tab);
-
-
-/****************************************************************************
-**
-*F  InitHasFiltListTNumsFromTable( <tab> )  . . initialise tester filts tnums
-*/
-extern void InitHasFiltListTNumsFromTable(const Int * tab);
-
-
-/****************************************************************************
-**
-*F  InitSetFiltListTNumsFromTable( <tab> )  . . initialise setter filts tnums
-*/
-extern void InitSetFiltListTNumsFromTable(const Int * tab);
-
-
-/****************************************************************************
-**
-*F  InitResetFiltListTNumsFromTable( <tab> )  initialise unsetter filts tnums
-*/
-extern void InitResetFiltListTNumsFromTable(const Int * tab);
-
-
-/****************************************************************************
-**
-*F  InitGVarFiltsFromTable( <tab> ) . . . . . . . . . . . . . . . new filters
-*/
-extern void InitGVarFiltsFromTable(const StructGVarFilt * tab);
-
-
-/****************************************************************************
-**
-*F  InitGVarAttrsFromTable( <tab> ) . . . . . . . . . . . . .  new attributes
-*/
-extern void InitGVarAttrsFromTable(const StructGVarAttr * tab);
-
-
-/****************************************************************************
-**
-*F  InitGVarPropsFromTable( <tab> ) . . . . . . . . . . . . .  new properties
-*/
-extern void InitGVarPropsFromTable(const StructGVarProp * tab);
-
-
-/****************************************************************************
-**
-*F  InitGVarOpersFromTable( <tab> ) . . . . . . . . . . . . .  new operations
-*/
-extern void InitGVarOpersFromTable(const StructGVarOper * tab);
-
-
-/****************************************************************************
-**
-*F  InitGVarFuncsFromTable( <tab> ) . . . . . . . . . . . . . .  new function
-*/
-extern void InitGVarFuncsFromTable(const StructGVarFunc * tab);
-
-
-/****************************************************************************
-**
-*F  InitHdlrFiltsFromTable( <tab> ) . . . . . . . . . . . . . . . new filters
-*/
-extern void InitHdlrFiltsFromTable(const StructGVarFilt * tab);
-
-
-/****************************************************************************
-**
-*F  InitHdlrAttrsFromTable( <tab> ) . . . . . . . . . . . . .  new attributes
-*/
-extern void InitHdlrAttrsFromTable(const StructGVarAttr * tab);
-
-
-/****************************************************************************
-**
-*F  InitHdlrPropsFromTable( <tab> ) . . . . . . . . . . . . .  new properties
-*/
-extern void InitHdlrPropsFromTable(const StructGVarProp * tab);
-
-
-/****************************************************************************
-**
-*F  InitHdlrOpersFromTable( <tab> ) . . . . . . . . . . . . .  new operations
-*/
-extern void InitHdlrOpersFromTable(const StructGVarOper * tab);
-
-
-/****************************************************************************
-**
-*F  InitHdlrFuncsFromTable( <tab> ) . . . . . . . . . . . . . . new functions
-*/
-extern void InitHdlrFuncsFromTable(const StructGVarFunc * tab);
-
-
-/****************************************************************************
-**
-*F  ImportGVarFromLibrary( <name>, <address> )  . . .  import global variable
-*/
-extern void ImportGVarFromLibrary(
-            const Char *        name,
-            Obj *               address );
-
-
-/****************************************************************************
-**
-*F  ImportFuncFromLibrary( <name>, <address> )  . . .  import global function
-*/
-extern void ImportFuncFromLibrary(
-            const Char *        name,
-            Obj *               address );
-
-
-/****************************************************************************
-**
-*F  Modules . . . . . . . . . . . . . . . . . . . . . . . . . list of modules
-*/
-typedef struct {
-
-    // pointer to the actual StructInitInfo
-    StructInitInfo * info;
-
-    // filename relative to GAP_ROOT or absolute
-    Char *           filename;
-
-    // true if the filename is GAP_ROOT relative
-    Int              isGapRootRelative;
-
-} StructInitInfoExt;
-
-
-extern void SaveModules(void);
-extern void LoadModules(void);
-
-extern Int ModulesPreSave(void);
-extern Int ModulesPostSave(void);
-
-
-/****************************************************************************
-**
-*F  RecordLoadedModule( <module> )  . . . . . . . . store module in <Modules>
-**
-**  The filename argument is a C string. A copy of it is taken in some
-**   private space and added to the module info.
-*/
-extern void RecordLoadedModule (
-    StructInitInfo *        module,
-    Int                     isGapRootRelative,
-    const Char *            filename );
-
-
 
 
 /****************************************************************************
