+++ src/system.h
@@ -818,206 +818,9 @@ extern UInt * * * SyAllocBags (
 extern void SyAbortBags(const Char * msg) NORETURN;
 
 
-/****************************************************************************
-**
-*F * * * * * * * * * * * * * loading of modules * * * * * * * * * * * * * * *
-**
-*/
-
-/****************************************************************************
-**
-*V  GAP_KERNEL_API_VERSION
-**
-**  'GAP_KERNEL_API_VERSION' gives the version of the GAP kernel. This value
-**  is used to check if kernel modules were built with a compatible kernel.
-**  This version is not the same as, and not connected to, the GAP version.
-**
-**  This is stored as
-**  'GAP_KERNEL_MAJOR_VERSION*1000 + GAP_KERNEL_MINOR_VERSION'.
-**
-**  The algorithm used is the following:
-**
-**  The kernel will not load a module compiled for a newer kernel.
-**
-**  The kernel will not load a module compiled for a different major version.
-**
-**  The minor version should be incremented when new backwards-compatible
-**  functionality is added. The major version should be incremented when
-**  a backwards-incompatible change is made.
-**
-**  The kernel version is a macro so it can be used by packages
-**  to optionally compile support for new functionality.
-**
-*/
-
-#define GAP_KERNEL_MAJOR_VERSION 1
-#define GAP_KERNEL_MINOR_VERSION 1
-#define GAP_KERNEL_API_VERSION                                               \
-    ((GAP_KERNEL_MAJOR_VERSION)*1000 + (GAP_KERNEL_MINOR_VERSION))
-
-enum {
-    /** builtin module */
-    MODULE_BUILTIN = GAP_KERNEL_API_VERSION * 10,
-
-    /** statically loaded compiled module */
-    MODULE_STATIC = GAP_KERNEL_API_VERSION * 10 + 1,
-
-    /** dynamically loaded compiled module */
-    MODULE_DYNAMIC = GAP_KERNEL_API_VERSION * 10 + 2,
-};
-
-static inline Int IS_MODULE_BUILTIN(UInt type)
-{
-    return type % 10 == 0;
-}
-
-static inline Int IS_MODULE_STATIC(UInt type)
-{
-    return type % 10 == 1;
-}
-
-static inline Int IS_MODULE_DYNAMIC(UInt type)
-{
-    return type % 10 == 2;
-}
-
-
-/****************************************************************************
-**
-*T  StructInitInfo  . . . . . . . . . . . . . . . . . module init information
-*/
-typedef struct init_info {
-
-    /* type of the module: MODULE_BUILTIN, MODULE_STATIC, MODULE_DYNAMIC   */
-    UInt             type;               
-
-    /* name of the module: filename with ".c" or library filename          */
-    const Char *     name;
-
-    /* revision entry of c file for MODULE_BUILTIN                         */
-    const Char *     revision_c;
-
-    /* revision entry of h file for MODULE_BUILTIN                         */
-    const Char *     revision_h;
-
-    /* version number for MODULE_BUILTIN                                   */
-    UInt             version;
-
-    /* CRC value for MODULE_STATIC or MODULE_DYNAMIC                       */
-    Int              crc;
-
-    /* initialise kernel data structures                                   */
-    Int              (* initKernel)(struct init_info *);
-
-    /* initialise library data structures                                  */
-    Int              (* initLibrary)(struct init_info *);
-
-    /* sanity check                                                        */
-    Int              (* checkInit)(struct init_info *);
-
-    /* function to call before saving workspace                            */
-    Int              (* preSave)(struct init_info *);
-
-    /* function to call after saving workspace                             */
-    Int              (* postSave)(struct init_info *);
-
-    /* function to call after restoring workspace                          */
-    Int              (* postRestore)(struct init_info *);
-
-} StructInitInfo;
-
+typedef struct init_info StructInitInfo;
 typedef StructInitInfo* (*InitInfoFunc)(void);
 
-
-/****************************************************************************
-**
-*T  StructBagNames  . . . . . . . . . . . . . . . . . . . . . tnums and names
-*/
-typedef struct {
-    Int             tnum;
-    const Char *    name;
-} StructBagNames;
-
-
-/****************************************************************************
-**
-*T  StructGVarFilt  . . . . . . . . . . . . . . . . . . . . . exported filter
-*/
-typedef struct {
-    const Char *    name;
-    const Char *    argument;
-    Obj *           filter;
-    Obj             (* handler)(/*arguments*/);
-    const Char *    cookie;
-} StructGVarFilt;
-
-// GVAR_FILTER a helper macro for quickly creating table entries in
-// StructGVarFilt, StructGVarAttr and StructGVarProp arrays
-#define GVAR_FILTER(name, argument, filter) \
-  { #name, argument, filter, Func ## name, __FILE__ ":" #name }
-
-
-/****************************************************************************
-**
-*T  StructGVarAttr  . . . . . . . . . . . . . . . . . . .  exported attribute
-*/
-typedef struct {
-    const Char *    name;
-    const Char *    argument;
-    Obj *           attribute;
-    Obj             (* handler)(/*arguments*/);
-    const Char *    cookie;
-} StructGVarAttr;
-
-
-/****************************************************************************
-**
-*T  StructGVarProp  . . . . . . . . . . . . . . . . . . . . exported property
-*/
-typedef struct {
-    const Char *    name;
-    const Char *    argument;
-    Obj *           property;
-    Obj             (* handler)(/*arguments*/);
-    const Char *    cookie;
-} StructGVarProp;
-
-
-/****************************************************************************
-**
-*T  StructGVarOper  . . . . . . . . . . . . . . . . . . .  exported operation
-*/
-typedef struct {
-    const Char *    name;
-    Int             nargs;
-    const Char *    args;
-    Obj *           operation;
-    Obj             (* handler)(/*arguments*/);
-    const Char *    cookie;
-} StructGVarOper;
-
-// GVAR_OPER is a helper macro for quickly creating table entries in
-// StructGVarOper arrays
-#define GVAR_OPER(name, nargs, args, operation) \
-  { #name, nargs, args, operation, Func ## name, __FILE__ ":" #name }
-
-
-/****************************************************************************
-**
-*T  StructGVarFunc  . . . . . . . . . . . . . . . . . . . . exported function
-*/
-typedef struct {
-    const Char *    name;
-    Int             nargs;
-    const Char *    args;
-    Obj             (* handler)(/*arguments*/);
-    const Char *    cookie;
-} StructGVarFunc;
-
-// GVAR_FUNC is a helper macro for quickly creating table entries in
-// StructGVarFunc arrays
-#define GVAR_FUNC(name, nargs, args) \
-  { #name, nargs, args, Func ## name, __FILE__ ":" #name }
 
 /****************************************************************************
 **
